+++ src/dto/productDTO.js
@@ -4,6 +4,7 @@ class ProductDto {
       this.description = product.description ?? 'No description';
       this.code = product.code ?? 'Unknown';
       this.price = product.price ?? null;
+      this.status = product.status ? "Available" : "Not Available";
       this.stock = product.stock ?? 0;
       this.category = product.category ?? 'Unknown';
       this.thumbnail = product.thumbnail ?? null;